Refactor suggestion picker
