+++ src/panels/lovelace/editor/card-editor/hui-suggestion-card.ts
@@ -10,7 +10,7 @@ import type { CardSuggestion } from "../../card-suggestions/types";
 
 declare global {
   interface HASSDomEvents {
-    "suggestion-picked": { suggestion: CardSuggestion };
+    "pick-suggestion": { suggestion: CardSuggestion };
   }
 }
 
@@ -21,6 +21,8 @@ const getItemCount = (config: LovelaceCardConfig): number => {
   return c.cards?.length ?? c.entities?.length ?? 0;
 };
 
+// Truncates preview content so the card grid stays responsive — full content
+// is restored when the suggestion is actually picked.
 const cappedCardConfig = (config: LovelaceCardConfig): LovelaceCardConfig => {
   const c = config as { cards?: LovelaceCardConfig[]; entities?: unknown[] };
   if (c.cards && c.cards.length > PREVIEW_ITEM_CAP) {
@@ -32,8 +34,8 @@ const cappedCardConfig = (config: LovelaceCardConfig): LovelaceCardConfig => {
   return config;
 };
 
-@customElement("hui-recipe-suggestion")
-export class HuiRecipeSuggestion extends LitElement {
+@customElement("hui-suggestion-card")
+export class HuiSuggestionCard extends LitElement {
   @property({ attribute: false }) public hass?: HomeAssistant;
 
   @property({ attribute: false }) public suggestion!: CardSuggestion;
@@ -50,28 +52,32 @@ export class HuiRecipeSuggestion extends LitElement {
       : undefined;
   }
 
-  private _renderPreview(): TemplateResult | typeof nothing {
-    if (!this._preview) return nothing;
-    return html`
-      <hui-card .hass=${this.hass} .config=${this._preview} preview></hui-card>
-    `;
-  }
-
   protected render(): TemplateResult {
     const { suggestion } = this;
     const totalCount = getItemCount(suggestion.config);
     const hiddenCount = Math.max(0, totalCount - PREVIEW_ITEM_CAP);
 
     return html`
-      <div class="card" tabindex="0">
-        <div
-          class="overlay"
-          role="button"
-          aria-label=${suggestion.label}
-          @click=${this._handleClick}
-        ></div>
+      <div
+        class="card"
+        tabindex="0"
+        role="button"
+        aria-label=${suggestion.label}
+        @click=${this._handleClick}
+        @keydown=${this._handleKeyDown}
+      >
         <div class="card-header">${suggestion.label}</div>
-        <div class="preview">${this._renderPreview()}</div>
+        <div class="preview">
+          ${this._preview
+            ? html`
+                <hui-card
+                  .hass=${this.hass}
+                  .config=${this._preview}
+                  preview
+                ></hui-card>
+              `
+            : nothing}
+        </div>
         ${hiddenCount > 0
           ? html`
               <div class="more-badge">
@@ -88,7 +94,14 @@ export class HuiRecipeSuggestion extends LitElement {
   }
 
   private _handleClick(): void {
-    fireEvent(this, "suggestion-picked", { suggestion: this.suggestion });
+    fireEvent(this, "pick-suggestion", { suggestion: this.suggestion });
+  }
+
+  private _handleKeyDown(ev: KeyboardEvent): void {
+    if (ev.key === "Enter" || ev.key === " ") {
+      ev.preventDefault();
+      this._handleClick();
+    }
   }
 
   static readonly styles: CSSResultGroup = css`
@@ -107,6 +120,10 @@ export class HuiRecipeSuggestion extends LitElement {
       overflow: hidden;
       border: var(--ha-card-border-width, 1px) solid
         var(--ha-card-border-color, var(--divider-color));
+    }
+    .card:focus-visible {
+      outline: 2px solid var(--primary-color);
+      outline-offset: 2px;
     }
     .card-header {
       color: var(--ha-card-header-color, var(--primary-text-color));
@@ -128,14 +145,6 @@ export class HuiRecipeSuggestion extends LitElement {
       display: block;
       width: 100%;
     }
-    .overlay {
-      position: absolute;
-      width: 100%;
-      height: 100%;
-      z-index: 1;
-      box-sizing: border-box;
-      border-radius: var(--ha-card-border-radius, var(--ha-border-radius-lg));
-    }
     .more-badge {
       margin: 0 var(--ha-space-4) var(--ha-space-3);
       padding: var(--ha-space-1) var(--ha-space-2);
@@ -151,6 +160,6 @@ export class HuiRecipeSuggestion extends LitElement {
 
 declare global {
   interface HTMLElementTagNameMap {
-    "hui-recipe-suggestion": HuiRecipeSuggestion;
+    "hui-suggestion-card": HuiSuggestionCard;
   }
 }
